2.1 Refactoring names, simplify 404, single style.
