+++ src/components/main/main.jsx
@@ -1,35 +1,14 @@
 import React from "react";
 import PropTypes from 'prop-types';
-import RoomCard from "../room-card/room-card";
-import {Link} from 'react-router-dom';
+import OfferCard from "../offer-card/offer-card";
+import Header from "../header/header";
 
 const Main = (props) => {
   const {roomQuantity} = props;
   const rooms = new Array(roomQuantity).fill();
   return (
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link header__logo-link--active" to="/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <Header/>
 
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
@@ -90,7 +69,7 @@ const Main = (props) => {
                 </ul>
               </form>
               <div className="cities__places-list places__list tabs__content">
-                {rooms.map((_, i) => <RoomCard key = {i}/>)}
+                {rooms.map((_, i) => <OfferCard key = {i}/>)}
               </div>
             </section>
             <div className="cities__right-section">
