--- src/components/page-not-found/page-not-found.jsx
@@ -1,48 +1,0 @@
-import React from "react";
-import {Link} from 'react-router-dom';
-
-const PageNotFound = () => {
-  return (
-    <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to="/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
-      <main>
-        <h1 style={{
-          textAlign: `center`,
-          color: `#4481c3`,
-          marginTop: `100px`,
-          marginBottom: `100px`,
-        }}>404 Page Not Found</h1>
-      </main>
-
-      <footer className="footer container">
-        <Link className="footer__logo-link" to="/">
-          <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33"/>
-        </Link>
-      </footer>
-    </div>
-  );
-};
-
-export default PageNotFound;
